chore: Rename files and imports for casing (#5)
