+++ pages/index.js
@@ -1,11 +1,11 @@
 import Link from "next/link";
-import { getPosts } from "../utils/mdxUtils";
+import { getPosts } from "../utils/mdx-utils";
 
 import Footer from "../components/Footer";
 import Header from "../components/Header";
 import Layout, { GradientBackground } from "../components/Layout";
 import ArrowIcon from "../components/ArrowIcon";
-import { getGlobalData } from "../utils/globalData";
+import { getGlobalData } from "../utils/global-data";
 import SEO from "../components/SEO";
 
 export default function Index({ posts, globalData }) {
