+++ pages/posts/[slug].js
@@ -1,10 +1,10 @@
-import { getGlobalData } from "../../utils/globalData";
+import { getGlobalData } from "../../utils/global-data";
 import {
   getNextPostBySlug,
   getPostBySlug,
   getPreviousPostBySlug,
   postFilePaths,
-} from "../../utils/mdxUtils";
+} from "../../utils/mdx-utils";
 
 import { MDXRemote } from "next-mdx-remote";
 import Head from "next/head";
