+++ utils/global-data.js
